bind: update named.conf
First, change the "directory" to point to /var/cache/bind, which ensures that
e.g. simple file "something" statements for secondary servers will create files
there (the directory is created by the init script, so we're sure that it
exists). This behaviour matches e.g. what Debian does.

Second, remove the "auth-nxdomain" stanza which is the default in bind since
9.0 was released.

Third, change "master" to "primary" (the wording used in the bind reference
docs)

Signed-off-by: David Härdeman <david@hardeman.nu>